document-qa-rag-full/backend/requirements.txt
fastapi
uvicorn[standard]
chromadb
python-multipart
pdfminer.six
python-docx
textract
google-genai
requests
